fix: add _libhoop/redactor/types subdirectory for package redactortypes
The stub _libhoop had guardrails types in redactor/types.go (package redactor),
but agent/controller imports redactortypes 'libhoop/redactor/types' which requires
a types/ subdirectory with package redactortypes. This caused CI build failures:
'module libhoop does not contain package libhoop/redactor/types'
